Feat: Creator Package Mngmt (#104)
* chore: aded new models

* chore: updated commission model

* chore: updated models

* chore: extracted creator creation to its own service to bettter handle their scenarios

* chore: added base wallet deposit/withdrawal for subscription

* chore: added wallet transaction model

* chore: fix lint

* chore: added subscription logic

* chore: added subscription to creator

* chore: added cancel subscription logic

* chore:
- added subscription activation endpoint
- query to get subs due for renewal

* chore: tested and fixed scenarios

* chore: added pricing activation

* chore: tested and fixed activate creator-subscriptions

+++ services/main/DTOs/OutputContent.cs
@@ -12,7 +12,7 @@ public class OutputContent
     public DateTime? DoneAt { get; set; } = null!;
     public List<OutputTag>? Tags { get; set; }
     public string Media { get; set; } = null!;
-    public int Amount { get; set; } = 0;
+    public Int64 Amount { get; set; } = 0;
     public required string CreatedById { get; set; }
     public OutputUser? CreatedBy { get; set; }
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
